Try django-parler/requirements.txt
Django==5.2.7
django-parler==2.3
